fix struct in one location
